app/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
APScheduler==3.7.0
astroid==2.4.2
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
coverage==5.3
cryptography==3.2.1
dnspython==2.0.0
docopt==0.6.2
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.61.2
flake8==3.8.4
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
metomi-isodatetime==1!2.0.2
orjson==3.4.4
passlib==1.7.4
promise==2.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.2
pyflakes==2.2.0
pylint==2.6.0
python-http-client==3.3.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
Rx==1.6.1
sendgrid==6.5.0
six==1.15.0
starkbank-ecdsa==1.1.0
starlette==0.13.6
toml==0.10.2
tzlocal==2.1
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
